tests/wpt/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/border-image-slice.html

<!doctype html>
<meta charset="utf-8">
<title>'border-image-slice' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

runPropertyTests('border-image-slice', [
  // Computed value is always four values, which are not supported in
  // Typed OM level 1.
  {
    syntax: '<number>',
    specified: assert_is_equal_with_range_handling,
    computed: (_, result) => assert_is_unsupported(result)
  },
  {
    syntax: '<percentage>',
    specified: assert_is_equal_with_range_handling,
    computed: (_, result) => assert_is_unsupported(result)
  },
]);

runUnsupportedPropertyTests('border-image-slice', [
  '30 fill', '30 40 50', '30 40 50 60 fill'
]);

</script>
